backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
Automat==24.8.1
backports.tarfile==1.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
configobj==5.0.8
constantly==23.10.4
cryptography==43.0.1
distro==1.9.0
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
entrypoints==0.4
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
hyperlink==21.0.0
idna==3.8
importlib_metadata==8.4.0
incremental==24.7.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jeepney==0.8.0
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyboard==0.13.5
keyring==25.3.0
launchpadlib==2.0.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
mailchimp-marketing==3.0.80
MarkupSafe==2.1.5
more-itertools==10.5.0
netifaces==0.11.0
oauthlib==3.2.2
openai==1.44.0
pbr==6.1.0
pexpect==4.9.0
pillow==10.4.0
psycopg==3.2.1
psycopg-binary==3.2.1
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
PyHamcrest==2.1.0
PyJWT==2.9.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
pyrsistent==0.20.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-unixsocket==0.3.0
rpds-py==0.20.0
SecretStorage==3.3.3
service-identity==24.1.0
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
ssh-import-id==5.11
testresources==2.0.1
tomli==2.0.1
tqdm==4.66.5
Twisted==24.7.0
typing_extensions==4.12.2
urllib3==2.2.2
wadllib==1.3.6
zipp==3.20.1
zope.interface==7.0.3
